add initial setup for logger library

+++ src/controllers/user.js
@@ -1,27 +1,34 @@
+const { ObjectId } = require('mongoose').Types;
 const UserModel = require('../models/user');
 
 module.exports = {
   getAllUsers: async (req, res) => {
     try {
       const users = await UserModel.find();
+      if (users.length <= 0)
+        throw new Error('There are no users at the moment!');
       res.json(users);
     } catch (err) {
-      res.status(422).json({ message: err.message });
+      res.status(422).json({ message: err.message ?? err });
     }
   },
   getSingleUser: async (req, res) => {
     const { id } = req.params;
     try {
+      if (String(new ObjectId(id)) !== id.toString())
+        throw new Error('Requested user ID is not valid!');
       const user = await UserModel.findById(id);
       if (!user) throw new Error("The user with the specified ID wasn't found");
       res.json(user);
     } catch (err) {
-      res.status(422).json({ message: err.message });
+      res.status(422).json({ message: err.message ?? err });
     }
   },
   updateUser: async (req, res) => {
     const { id } = req.params;
     try {
+      if (String(new ObjectId(id)) !== id.toString())
+        throw new Error('Requested user ID is not valid!');
       const updatedUser = await UserModel.findByIdAndUpdate(
         id,
         {
@@ -37,7 +44,7 @@ module.exports = {
 
       res.json(updatedUser);
     } catch (err) {
-      res.status(422).json({ message: err.message });
+      res.status(422).json({ message: err.message ?? err });
     }
   },
 
@@ -45,10 +52,14 @@ module.exports = {
   deleteUser: async (req, res) => {
     const { id } = req.params;
     try {
-      await UserModel.findByIdAndRemove(id);
+      if (String(new ObjectId(id)) !== id.toString())
+        throw new Error('Requested user ID is not valid!');
+      const deletedUser = await UserModel.findByIdAndRemove(id);
+      if (!deletedUser)
+        throw new Error("The user with the specified ID wasn't found");
       res.status(204).end();
     } catch (err) {
-      res.status(422).json({ message: err.message });
+      res.status(422).json({ message: err.message ?? err });
     }
   },
   getAllDonators: async (req, res) => {
